Update project files for XCode 4.6 on Mac OS 10.8

+++ FlySightViewer/Mac/CorePlot/framework/Source/CPGradient.m
@@ -663,7 +663,7 @@ static void resolveHSV(CGFloat *color1, CGFloat *color2);
     CPGradientElement removedElement = [newGradient removeElementAtIndex:index];
 	
     if ( isnan(removedElement.position) ) {
-		[NSException raise:NSRangeException format:@"-[%@ removeColorStopAtIndex:]: index (%i) beyond bounds", [self class], index];
+		[NSException raise:NSRangeException format:@"-[%@ removeColorStopAtIndex:]: index (%li) beyond bounds", [self class], (unsigned long)index];
 	}
 	
     return [newGradient autorelease];
@@ -690,7 +690,7 @@ static void resolveHSV(CGFloat *color1, CGFloat *color2);
 		
 	}
 	
-    [NSException raise:NSRangeException format:@"-[%@ colorStopAtIndex:]: index (%i) beyond bounds", [self class], index];
+    [NSException raise:NSRangeException format:@"-[%@ colorStopAtIndex:]: index (%li) beyond bounds", [self class], (unsigned long)index];
 	
     return NULL;
 }
